:art: Move find_template_match and helping fns to new file
